Extract decoding of user id to lambda layer
