+++ backend/src/lambdas/add-history-record/package.json
@@ -4,10 +4,6 @@
   "main": "index.mjs",
   "type": "module",
   "dependencies": {
-    "@aws-sdk/util-dynamodb": "^3.787.0",
-    "jsonwebtoken": "^9.0.2"
-  },
-  "devDependencies": {
-    "@types/jsonwebtoken": "^9.0.9"
+    "@aws-sdk/util-dynamodb": "^3.787.0"
   }
 }
