+++ backend/src/lambdas/get-character/index.ts
@@ -1,6 +1,5 @@
 import { APIGatewayProxyEvent, APIGatewayProxyResult } from "aws-lambda";
-import jwt, { JwtPayload } from "jsonwebtoken";
-import { Request, parseBody, getCharacterItem, HttpError, ensureHttpError } from "utils/index.js";
+import { Request, parseBody, getCharacterItem, HttpError, ensureHttpError, decodeUserId } from "utils/index.js";
 
 export const handler = async (event: APIGatewayProxyEvent): Promise<APIGatewayProxyResult> => {
   return getCharacter({
@@ -36,23 +35,7 @@ export async function getCharacter(request: Request): Promise<APIGatewayProxyRes
 function validateRequest(request: Request): Parameters {
   console.log("Validate request");
 
-  // Trim the authorization header as it could contain spaces at the beginning
-  const authHeader = request.headers.Authorization?.trim() || request.headers.authorization?.trim();
-  if (!authHeader || !authHeader.startsWith("Bearer ")) {
-    throw new HttpError(401, "Unauthorized: No token provided!");
-  }
-
-  const token = authHeader.split(" ")[1]; // Remove "Bearer " prefix
-  // Decode the token without verification (the access to the API itself is already protected by the authorizer)
-  const decoded = jwt.decode(token) as JwtPayload | null;
-  if (!decoded) {
-    throw new HttpError(401, "Unauthorized: Invalid token!");
-  }
-
-  const userId = decoded.sub; // Cognito User ID
-  if (!userId) {
-    throw new HttpError(401, "Unauthorized: User ID not found in token!");
-  }
+  const userId = decodeUserId(request.headers.Authorization);
 
   if (typeof request.pathParameters?.["character-id"] !== "string") {
     throw new HttpError(400, "Invalid input values!");
