+++ backend/src/lambdas/get-skill-increase-cost/index.ts
@@ -1,7 +1,6 @@
 import { APIGatewayProxyEvent, APIGatewayProxyResult } from "aws-lambda";
 import { adjustCostCategory, Character, getSkillIncreaseCost, getSkill, parseLearningMethod } from "config/index.js";
-import jwt, { JwtPayload } from "jsonwebtoken";
-import { Request, parseBody, getCharacterItem } from "utils/index.js";
+import { Request, parseBody, getCharacterItem, decodeUserId } from "utils/index.js";
 
 export const handler = async (event: APIGatewayProxyEvent): Promise<APIGatewayProxyResult> => {
   return getSkillCost({
@@ -70,32 +69,7 @@ export async function getSkillCost(request: Request): Promise<APIGatewayProxyRes
 function validateRequest(request: Request): Parameters {
   console.log("Validate request");
 
-  // Trim the authorization header as it could contain spaces at the beginning
-  const authHeader = request.headers.Authorization?.trim() || request.headers.authorization?.trim();
-  if (!authHeader || !authHeader.startsWith("Bearer ")) {
-    throw {
-      statusCode: 401,
-      body: JSON.stringify({ message: "Unauthorized: No token provided!" }),
-    };
-  }
-
-  const token = authHeader.split(" ")[1]; // Remove "Bearer " prefix
-  // Decode the token without verification (the access to the API itself is already protected by the authorizer)
-  const decoded = jwt.decode(token) as JwtPayload | null;
-  if (!decoded) {
-    throw {
-      statusCode: 401,
-      body: JSON.stringify({ message: "Unauthorized: Invalid token!" }),
-    };
-  }
-
-  const userId = decoded.sub; // Cognito User ID
-  if (!userId) {
-    throw {
-      statusCode: 401,
-      body: JSON.stringify({ message: "Unauthorized: User ID not found in token!" }),
-    };
-  }
+  const userId = decodeUserId(request.headers.Authorization);
 
   if (
     typeof request.pathParameters?.["character-id"] !== "string" ||
