+++ backend/src/lambdas/get-skill-increase-cost/package.json
@@ -3,12 +3,8 @@
   "version": "1.0.0",
   "main": "index.mjs",
   "type": "module",
-  "dependencies": {
-    "jsonwebtoken": "^9.0.2"
-  },
   "devDependencies": {
     "@types/aws-lambda": "^8.10.145",
-    "@types/jsonwebtoken": "^9.0.9",
     "@types/node": "^22.9.0"
   }
 }
