+++ backend/src/lambdas/increase-skill/index.ts
@@ -1,5 +1,4 @@
 import { APIGatewayProxyEvent, APIGatewayProxyResult } from "aws-lambda";
-import jwt, { JwtPayload } from "jsonwebtoken";
 import {
   parseLearningMethod,
   adjustCostCategory,
@@ -8,7 +7,7 @@ import {
   getSkill,
   Skill,
 } from "config/index.js";
-import { Request, parseBody, getCharacterItem, updateSkill } from "utils/index.js";
+import { Request, parseBody, getCharacterItem, updateSkill, decodeUserId } from "utils/index.js";
 
 export const handler = async (event: APIGatewayProxyEvent): Promise<APIGatewayProxyResult> => {
   return increaseSkill({
@@ -142,32 +141,7 @@ export async function increaseSkill(request: Request): Promise<APIGatewayProxyRe
 function validateRequest(request: Request): Parameters {
   console.log("Validate request");
 
-  // Trim the authorization header as it could contain spaces at the beginning
-  const authHeader = request.headers.Authorization?.trim() || request.headers.authorization?.trim();
-  if (!authHeader || !authHeader.startsWith("Bearer ")) {
-    throw {
-      statusCode: 401,
-      body: JSON.stringify({ message: "Unauthorized: No token provided!" }),
-    };
-  }
-
-  const token = authHeader.split(" ")[1]; // Remove "Bearer " prefix
-  // Decode the token without verification (the access to the API itself is already protected by the authorizer)
-  const decoded = jwt.decode(token) as JwtPayload | null;
-  if (!decoded) {
-    throw {
-      statusCode: 401,
-      body: JSON.stringify({ message: "Unauthorized: Invalid token!" }),
-    };
-  }
-
-  const userId = decoded.sub; // Cognito User ID
-  if (!userId) {
-    throw {
-      statusCode: 401,
-      body: JSON.stringify({ message: "Unauthorized: User ID not found in token!" }),
-    };
-  }
+  const userId = decodeUserId(request.headers.Authorization);
 
   /**
    * This conversion is necessary if the Lambda is called via AWS Step Functions.
