+++ backend/src/utils/index.ts
@@ -2,3 +2,4 @@ export { Request, parseBody } from "./request.js";
 export { getCharacterItem, getCharacterItems, updateSkill } from "./dynamodb_characters.js";
 export { getHistoryItems, createHistoryItem, addHistoryRecord } from "./dynamodb_history.js";
 export { ensureHttpError, HttpError } from "./errors.js";
+export { decodeUserId } from "./token.js";
