+++ backend/src/utils/package.json
@@ -6,6 +6,7 @@
   "dependencies": {
     "@aws-sdk/client-dynamodb": "^3.788.0",
     "@aws-sdk/lib-dynamodb": "^3.788.0",
+    "jsonwebtoken": "^9.0.2",
     "zod": "^3.24.2"
   }
 }
